A few spanish translations
About 783-825: I think these are the ones used for dates, so I left them
in proper form. It should be "el {daynumber} de {month}" when formatted.
(either abbreviated or not)
I'm not very good at my Spanish, so this is just a first run through of
things I know for sure. More to come soon.