JP: A bunch of miscellaneous additions
I went through at random and added a bunch of things that I felt I could translate with a reasonably high degree of confidence (since I'm not a native speaker). I'll try to consult with native speakers to translate some of the really tricky stuff if nobody else gets to it in the mean time.