de-DE: adjust translation of STR_6515
The sentence should not contain in upper case after the hyphen. Either it must continue in lower case or the hyphen can be changed to a full stop.
Argument for the former is that it is closer to the english sentence. Argument for the latter is that it is more consistent with STR_6516 and STR_6517.
A (subjective) choice was made for the latter.