en-US: Remove "Tunnel of Horror" for falling back to "Ghost Train"
This name has puzzled me for years, i've never once seen this ride type refered to as a "Tunnel of Horror" in real life and almost always see it refered to as a Ghost Train, Haunted Mansion Ride, or just simply Dark Ride. I've opted for removing it so that it falls back to "Ghost Train" from en-GB but i am open for the other two options i mentioned if those might make more sense in en-US vs Ghost Train.