Improve STR_5612 and STR_5961 in sv-SE
STR_5612: "Spök" literally means ghost, but it is not a word that is commonly used in the context of ghost elements in computer games. "Skiss" means sketch, draft or outline, which rather translates the idea of what it is.

STR_5961: "entré" means entry in the sense of entrance, but not in the sense of a list item.